requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
openai==1.3.0
SpeechRecognition==3.10.0
gTTS==2.3.2
python-dotenv==1.0.0
requests==2.31.0
Werkzeug==2.3.7
